Refactor UserController and update routes
Renamed the 'create' method to 'store' in UserController to adhere to Laravel naming conventions. Added email validation for the 'store' method in UserController. Enhanced the messaged returned when the 'show' method is called to provide clearer information. Additionally, routes for user-related functions have been nested under a 'users' prefix for improved organization and consistency.